refactor: 복잡성 감소를 위한 서비스 통합
