+++ src/main/java/roomescape/presentation/api/reservation/PersonalReservationController.java
@@ -4,7 +4,7 @@ import java.util.List;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.application.reservation.PersonalReservationService;
+import roomescape.application.reservation.ReservationService;
 import roomescape.application.reservation.dto.ReservationWithStatusResult;
 import roomescape.presentation.api.reservation.response.ReservationWithStatusResponse;
 import roomescape.presentation.support.methodresolver.AuthInfo;
@@ -13,15 +13,15 @@ import roomescape.presentation.support.methodresolver.AuthPrincipal;
 @RestController
 public class PersonalReservationController {
 
-    private final PersonalReservationService personalReservationService;
+    private final ReservationService reservationService;
 
-    public PersonalReservationController(PersonalReservationService personalReservationService) {
-        this.personalReservationService = personalReservationService;
+    public PersonalReservationController(ReservationService reservationService) {
+        this.reservationService = reservationService;
     }
 
     @GetMapping("reservations-mine")
     public ResponseEntity<List<ReservationWithStatusResponse>> findMineReservations(@AuthPrincipal AuthInfo authInfo) {
-        List<ReservationWithStatusResult> reservationsWithStatus = personalReservationService.findReservationsWithStatus(
+        List<ReservationWithStatusResult> reservationsWithStatus = reservationService.findReservationsWithStatus(
                 authInfo.memberId()
         );
         List<ReservationWithStatusResponse> reservationWithStatusResponses = reservationsWithStatus.stream()
