+++ src/test/java/roomescape/application/reservation/ReservationServiceTest.java
@@ -15,6 +15,7 @@ import roomescape.application.reservation.dto.CreateReservationParam;
 import roomescape.application.reservation.dto.ReservationResult;
 import roomescape.application.reservation.dto.ReservationSearchParam;
 import roomescape.application.reservation.dto.ReservationTimeResult;
+import roomescape.application.reservation.dto.ReservationWithStatusResult;
 import roomescape.application.reservation.dto.ThemeResult;
 import roomescape.application.support.exception.NotFoundEntityException;
 import roomescape.domain.BusinessRuleViolationException;
@@ -24,6 +25,7 @@ import roomescape.domain.member.MemberRepository;
 import roomescape.domain.member.Role;
 import roomescape.domain.reservation.Reservation;
 import roomescape.domain.reservation.ReservationRepository;
+import roomescape.domain.reservation.ReservationStatus;
 import roomescape.domain.reservation.ReservationTime;
 import roomescape.domain.reservation.ReservationTimeRepository;
 import roomescape.domain.reservation.Theme;
@@ -238,5 +240,40 @@ class ReservationServiceTest extends AbstractServiceIntegrationTest {
                                 )
                         )
                 );
+    }
+
+    @Test
+    void 사용자의_예약을_모두_조회할_수_있다() {
+        // given
+        Member member = memberRepository.save(new Member("벨로", new Email("test@email.com"), "pw", Role.NORMAL));
+        Theme theme = themeRepository.save(new Theme("테마", "설명", "이미지"));
+        ReservationTime time1 = reservationTimeRepository.save(new ReservationTime(LocalTime.of(13, 0)));
+        ReservationTime time2 = reservationTimeRepository.save(new ReservationTime(LocalTime.of(14, 0)));
+        reservationRepository.save(new Reservation(member, LocalDate.now(clock), time1, theme));
+        reservationRepository.save(new Reservation(member, LocalDate.now(clock).plusDays(1), time2, theme));
+
+        // when
+        List<ReservationWithStatusResult> reservationsWithStatus = reservationService.findReservationsWithStatus(
+                member.getId()
+        );
+
+        // then
+        assertThat(reservationsWithStatus)
+                .isEqualTo(List.of(
+                        new ReservationWithStatusResult(
+                                1L,
+                                "테마",
+                                LocalDate.now(clock),
+                                LocalTime.of(13, 0),
+                                ReservationStatus.RESERVE
+                        ),
+                        new ReservationWithStatusResult(
+                                2L,
+                                "테마",
+                                LocalDate.now(clock).plusDays(1),
+                                LocalTime.of(14, 0),
+                                ReservationStatus.RESERVE
+                        )
+                ));
     }
 }
